Fix one path of index module depending on search.
When writing colbert encoder, I found any complex or compositional
encoder has good chance to depend on index module (because such index
needs more complex RepresentationWriter), and if index module depends
on search module, it has good chance to depend on the encode module,
thus a loop of dependency is formed.

We can fix it by cut the very counterintuitive "index --> search"
dependency arrow because it violates the common abstraction.
