+++ pyserini/search/__init__.py
@@ -14,12 +14,12 @@
 # limitations under the License.
 #
 
-from ._base import Document, JDocument, JQuery, JDisjunctionMaxQueryGenerator, get_topics, get_topics_with_reader, get_qrels_file, get_qrels
+from ._base import JQuery, JDisjunctionMaxQueryGenerator, get_topics, get_topics_with_reader, get_qrels_file, get_qrels
 from ._searcher import JSimpleSearcherResult, LuceneSimilarities, SimpleFusionSearcher, SimpleSearcher
 from ._nearest_neighbor import SimpleNearestNeighborSearcher, JSimpleNearestNeighborSearcherResult
 from ._impact_searcher import JImpactSearcherResult, ImpactSearcher
 
-__all__ = ['Document', 'JDocument', 'JQuery', 'LuceneSimilarities', 'SimpleFusionSearcher', 'SimpleSearcher',
+__all__ = ['JQuery', 'LuceneSimilarities', 'SimpleFusionSearcher', 'SimpleSearcher',
            'JSimpleSearcherResult', 'SimpleNearestNeighborSearcher', 'JSimpleNearestNeighborSearcherResult',
            'ImpactSearcher', 'JImpactSearcherResult', 'JDisjunctionMaxQueryGenerator', 'get_topics',
            'get_topics_with_reader', 'get_qrels_file', 'get_qrels']
